requirements.txt
attrs==22.1.0
aws-cdk-lib==2.39.1
boto3==1.24.63
botocore==1.27.63
cattrs==22.1.0
cdk-dynamo-table-view==0.2.256
constructs==10.1.91
exceptiongroup==1.0.0rc9
jmespath==1.0.1
jsii==1.66.0
publication==0.0.3
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
typeguard==2.13.3
types-awscrt==0.14.3
types-s3transfer==0.6.0.post4
typing_extensions==4.3.0
urllib3==1.26.12
